separate context from outside
