+++ src/connection/server_support.c
@@ -2408,7 +2408,7 @@ css_server_task::execute (context_type &thread_ref)
 
   if (m_conn.end_working_task () == 0 && m_conn.status == CONN_CLOSING)
     {
-      css_request_shutdown_conn (&m_conn, static_cast <uint8_t> (cubconn::connection_worker::ignore_level::DONT_IGNORE), false, 0 /* no wait */);
+      css_request_shutdown_conn (&m_conn, static_cast <uint8_t> (cubconn::connection_worker_ignore::DONT_IGNORE), false, 0 /* no wait */);
     }
   else
     {
